+++ include/ER_OLEDM1_CH1115_graphics.h
@@ -2,7 +2,7 @@
 * Project Name: ER_OLEDM1_CH1115
 * File: ER_OLEDM1_CH1115_graphics.h
 * Description: ER_OLEDM1 OLED driven by CH1115 controller header file 
-* for the graphics functions based on Adafruit graphics library
+* for the graphics functions.
 * Author: Gavin Lyons.
 * URL: https://github.com/gavinlyonsrepo/ER_OLEDM1_CH1115_RPI
 */
@@ -15,6 +15,33 @@
 #include <stdint.h>
 
 #define swap(a, b) { int16_t t = a; a = b; b = t; }
+
+typedef enum 
+{
+    CH1115Font_Default = 1,
+    CH1115Font_Thick = 2,
+    CH1115Font_Seven_Seg = 3,
+    CH1115Font_Wide = 4,
+    CH1115Font_Bignum = 5,
+    CH1115Font_Mednum = 6,
+}OLED_FONT_TYPE_e;
+
+typedef enum 
+{
+	FONT_W_5 = 5, FONT_W_7 = 7, FONT_W_4 = 4, FONT_W_8 = 8,FONT_W_16= 16
+}OLED_Font_width_e; // width of the font in bytes cols.
+
+typedef enum 
+{
+	FONT_O_EXTEND = 0x00, //   extends ASCII
+	FONT_O_SP = 0x20,  // Starts at Space
+	FONT_O_NUM = 0x30,  // Starts at number '0'
+}OLED_Font_offset_e; // font offset in the ASCII table
+
+typedef enum 
+{
+	FONT_H_8 = 8, FONT_H_16 = 16, FONT_H_32 = 32
+}OLED_Font_height_e; // width of the font in bits
 
 class ERMCH1115_graphics : public Print{
 
@@ -55,14 +82,19 @@ class ERMCH1115_graphics : public Print{
 	setTextColor(uint8_t c, uint8_t bg),
 	setTextSize(uint8_t s),
 	setTextWrap(bool w);
+	
+	void setDrawBitmapAddr(bool mode);
+	void drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap,
+			int16_t w, int16_t h, uint8_t color, uint8_t bg);
 
 	virtual size_t write(uint8_t);
 
-	void SetFontNum(uint8_t FontNumber);
+	void setFontNum(OLED_FONT_TYPE_e FontNumber);
 	void drawCharNumFont(uint8_t x, uint8_t y, uint8_t c, uint8_t color , uint8_t bg);
 	void drawTextNumFont(uint8_t x, uint8_t y, char *pText, uint8_t color, uint8_t bg);
 	int16_t height(void) const;
 	int16_t width(void) const;
+	OLED_FONT_TYPE_e FontNum;
 
  protected:
 	const int16_t WIDTH;
@@ -77,10 +109,12 @@ class ERMCH1115_graphics : public Print{
 	uint8_t textsize;
 
 	bool wrap; // If set, 'wrap' text at right edge of display
-	uint8_t _FontNumber = 1;
-	uint8_t _CurrentFontWidth = 5;
-	uint8_t _CurrentFontoffset = 0;
-	uint8_t _CurrentFontheight = 8;
+	bool drawBitmapAddr; // True = vertical , false = horizontal
+	
+	uint8_t _FontNumber = CH1115Font_Default;;
+	uint8_t _CurrentFontWidth = FONT_W_5;
+	uint8_t _CurrentFontoffset = FONT_O_EXTEND ;
+	uint8_t _CurrentFontheight = FONT_H_8;
 };
 
 #endif