+++ src/ER_OLEDM1_CH1115.cpp
@@ -538,4 +538,20 @@ void ERMCH1115::drawPixel(int16_t x, int16_t y, uint8_t colour)
 	return;
 }
 
+// Func Desc: init the Multibuffer struct
+// Param 1 Pointer to a struct
+// Param 2 Pointer to buffer array data(arrays decay to  pointers)
+// Param 3. width of buffer
+// Param 4. height of buffer
+// Param 5. x offset of buffer
+// Param 6. y offset of buffer
+void ERMCH1115::OLEDinitBufferStruct(MultiBuffer* mystruct, uint8_t* mybuffer, uint8_t w,  uint8_t h, int16_t  x, int16_t y)
+{
+   mystruct->screenbitmap = mybuffer; // point it to the buffer
+   mystruct->width = w ;
+   mystruct->height = h;
+   mystruct->xoffset = x;
+   mystruct->yoffset = y; 
+}
+
 //***********************************************